// src/app/models/admin/dangkyhoatdongkhoangsan/dangkythamdo/dkthamdocongtrinh.model.ts
export class OutputDkThamDoCongTrinhModel {
  idcongtrinh: string;
  sohieu: string;
  chieusau: number;
  toadox: number;
  toadoy: number;
  iddangkythamdo: string;
  geom: string;
  geovn: string;
  geowgs: string;
  hequychieu: string;
  ghichu: string;
  siteid: string;
}

export class InputDkThamDoCongTrinhModel {
  idcongtrinh: string;
  sohieu: string;
  chieusau: number;
  toadox: number;
  toadoy: number;
  iddangkythamdo: string;
  geom: string;
  geovn: string;
  geowgs: string;
  hequychieu: string;
  ghichu: string;
  siteid: string;
}